Refactor for new Demo
