+++ Groundbreakers/Assets/Prefabs/Tiles/TileBlockA.prefab
@@ -38,7 +38,7 @@ SpriteRenderer:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 3120260381402315756}
-  m_Enabled: 1
+  m_Enabled: 0
   m_CastShadows: 0
   m_ReceiveShadows: 0
   m_DynamicOccludee: 1
@@ -94,7 +94,7 @@ GameObject:
   - component: {fileID: 7085035497353382218}
   - component: {fileID: 3860437930588547951}
   m_Layer: 0
-  m_Name: TileBlock
+  m_Name: TileBlockA
   m_TagString: Tile
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -196,7 +196,7 @@ MonoBehaviour:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 5300740713263289941}
-  m_Enabled: 1
+  m_Enabled: 0
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: eecfd56d319619548bcd63f0f5b7c36d, type: 3}
   m_Name: 
@@ -211,7 +211,7 @@ LineRenderer:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 5300740713263289941}
-  m_Enabled: 1
+  m_Enabled: 0
   m_CastShadows: 1
   m_ReceiveShadows: 1
   m_DynamicOccludee: 1
@@ -306,13 +306,16 @@ MonoBehaviour:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 5300740713263289941}
-  m_Enabled: 1
+  m_Enabled: 0
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 748a71bdf4580624ab46a9f92ea21425, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   characterOnTop: 0
   canDeploy: 1
+  deploying: 0
+  deployingCharacter: -1
   canDeployIcon: {fileID: 21300000, guid: b1c0eb309fbe52a458268b40bdd936f4, type: 3}
   canNotDeployIcon: {fileID: 21300000, guid: b969ae0e2c061b3438f61074a306dafc, type: 3}
   occupiedIcon: {fileID: 21300000, guid: 0c1d367553f98824fa62a179c9aa5eed, type: 3}
+  deploy: {fileID: 0}
