+++ Groundbreakers/Assets/Scripts/Core/NewBehaviourScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0af55c33202d01c4782787f5b1ae64bf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
