+++ Groundbreakers/Assets/Scripts/TileMaps/TileController.cs
@@ -1,0 +1,11 @@
+﻿namespace TileMaps
+{
+    using UnityEngine;
+
+    /// <summary>
+    ///     This component provide APIs to modify any tiles: swap, destruct, or construct.
+    /// </summary>
+    public class TileController : MonoBehaviour
+    {
+    }
+}
